Moving functions around with reckless abandon
